fix plant patch

+++ backend/nursery/src/main/java/com/nursery/nursery/dto/PlantDTO.java
@@ -14,7 +14,7 @@ public class PlantDTO {
     Long id;
     String name;
     String category;
-    Double plantHeight;
+    Float plantHeight;
     Integer plantsStock;
     Float cost;
 }
